+++ HTTP ECGridOS API v4.1 .NET 6 Console App/ECGridOSClient.cs
@@ -30,8 +30,8 @@ public interface IECGridOSClient
     Task<List<CallBackQueueIDInfo>> CallBackPendingList(string sessionID);
     Task<List<CallBackQueueIDInfo>> CallBackPendingListEx(string sessionID, int networkID, int mailboxID);
     Task<List<CallBackQueueIDInfo>> CallBackPendingListExA(string sessionID, int networkID, int mailboxID, string networkExclude);
-    Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, int callBackQueueID);
-    Task<CallBackQueueIDInfo> CallBackTest(string sessionID, int callBackQueueID, int parcelID, int interchangeID, int userID);
+    Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, long callBackQueueID);
+    Task<CallBackQueueIDInfo> CallBackTest(string sessionID, long callBackQueueID, long parcelID, long interchangeID, int userID);
 
     #endregion
 
@@ -94,7 +94,7 @@ public interface IECGridOSClient
 
     #region Interchange
 
-    Task<bool> InterchangeCancel(string sessionID, int interchangeID);
+    Task<bool> InterchangeCancel(string sessionID, long interchangeID);
     Task<DateTime> InterchangeDate(string sessionID, string interchangeHeader);
     Task<InterchangeIDInfo> InterchangeHeaderInfo(string sessionID, string interchangeHeader);
     Task<InterchangeIDInfo> InterchangeHeaderInfoB(string sessionID, byte[] interchangeHeader);
@@ -140,10 +140,10 @@ public interface IECGridOSClient
 
     #region Key Set/Get/List/Remove
 
-    Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared);
-    Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, int objectID);
-    Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared);
-    Task<bool> KeySet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility, string value, string meta, int daysToLive);
+    Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared);
+    Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, long objectID);
+    Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared);
+    Task<bool> KeySet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility, string value, string meta, int daysToLive);
 
     #endregion
 
@@ -240,15 +240,15 @@ public interface IECGridOSClient
     Task<ParcelIDInfoCollection> ParcelOutBoxInProcess(string sessionID);
     Task<ParcelIDInfoCollection> ParcelOutBoxInProcessEx(string sessionID, int networkID, int mailboxID);
     Task<bool> ParcelSetMailbagControlID(string sessionID, long parcelID, string mailbagControlID);
-    Task<int> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType);
+    Task<long> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType);
     Task<bool> ParcelUpdateLocalStatus(string sessionID, long parcelID, short status);
     Task<bool> ParcelUpdateStatus(string sessionID, long parcelID, ParcelStatus status, bool transLogOnly);
-    Task<int> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content);
-    Task<int> ParcelUploadA(string sessionID, string fileName, string contentBase64);
-    Task<int> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content);
-    Task<int> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64);
-    Task<int> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo);
-    Task<int> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo);
+    Task<long> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content);
+    Task<long> ParcelUploadA(string sessionID, string fileName, string contentBase64);
+    Task<long> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content);
+    Task<long> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64);
+    Task<long> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo);
+    Task<long> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo);
 
     #endregion
 
@@ -770,7 +770,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, int callBackQueueID)
+    public async Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, long callBackQueueID)
     {
         try
         {
@@ -811,7 +811,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<CallBackQueueIDInfo> CallBackTest(string sessionID, int callBackQueueID, int parcelID, int interchangeID, int userID)
+    public async Task<CallBackQueueIDInfo> CallBackTest(string sessionID, long callBackQueueID, long parcelID, long interchangeID, int userID)
     {
         try
         {
@@ -2158,7 +2158,7 @@ public class ECGridOSClient : IECGridOSClient
 
     #region Interchange
 
-    public async Task<bool> InterchangeCancel(string sessionID, int interchangeID)
+    public async Task<bool> InterchangeCancel(string sessionID, long interchangeID)
     {
         try
         {
@@ -3758,7 +3758,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="objectID"></param>
     /// <param name="visibility"></param>
     /// <returns>KeyValue</returns>
-    public async Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared)
+    public async Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared)
     {
         try
         {
@@ -3809,7 +3809,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="systemObject"></param>
     /// <param name="objectID"></param>
     /// <returns>List of KeyValue Objects</returns>
-    public async Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, int objectID)
+    public async Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, long objectID)
     {
         try
         {
@@ -3860,7 +3860,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="objectID"></param>
     /// <param name="visibility"></param>
     /// <returns>True/False</returns>
-    public async Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared)
+    public async Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared)
     {
         try
         {
@@ -3916,7 +3916,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="meta"></param>
     /// <param name="daysToLive"></param>
     /// <returns>True/False</returns>
-    public async Task<bool> KeySet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility, string value, string meta, int daysToLive)
+    public async Task<bool> KeySet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility, string value, string meta, int daysToLive)
     {
         try
         {
@@ -6898,7 +6898,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType)
+    public async Task<long> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType)
     {
         try
         {
@@ -6923,8 +6923,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7026,7 +7026,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content)
+    public async Task<long> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content)
     {
         try
         {
@@ -7051,8 +7051,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7069,7 +7069,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadA(string sessionID, string fileName, string contentBase64)
+    public async Task<long> ParcelUploadA(string sessionID, string fileName, string contentBase64)
     {
         try
         {
@@ -7093,8 +7093,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7111,7 +7111,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content)
+    public async Task<long> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content)
     {
         try
         {
@@ -7138,8 +7138,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7156,7 +7156,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64)
+    public async Task<long> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64)
     {
         try
         {
@@ -7182,8 +7182,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7200,7 +7200,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo)
+    public async Task<long> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo)
     {
         try
         {
@@ -7227,8 +7227,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7245,7 +7245,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo)
+    public async Task<long> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo)
     {
         try
         {
@@ -7271,8 +7271,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
