+++ HTTP ECGridOS API v4.1 .NET 6 Worker Service/ECGridOSClient.cs
@@ -32,8 +32,8 @@ public interface IECGridOSClient
     Task<List<CallBackQueueIDInfo>> CallBackPendingList(string sessionID);
     Task<List<CallBackQueueIDInfo>> CallBackPendingListEx(string sessionID, int networkID, int mailboxID);
     Task<List<CallBackQueueIDInfo>> CallBackPendingListExA(string sessionID, int networkID, int mailboxID, string networkExclude);
-    Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, int callBackQueueID);
-    Task<CallBackQueueIDInfo> CallBackTest(string sessionID, int callBackQueueID, int parcelID, int interchangeID, int userID);
+    Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, long callBackQueueID);
+    Task<CallBackQueueIDInfo> CallBackTest(string sessionID, long callBackQueueID, long parcelID, long interchangeID, int userID);
 
     #endregion
 
@@ -83,7 +83,7 @@ public interface IECGridOSClient
     Task<bool> CommSetStatus(string sessionID, int commID, Status status);
     Task<CommIDInfo> CommUpdate(string sessionID, int commID, int ownerUserID, string identifier, string uRL, string version, bool signData, bool encryptData, bool compressData, ReceiptType receiptType,
         HTTPAuthType hTTPAuthentication, string hTTPUser, string hTTPPassword, UseType useType, DateTime beginDate, DateTime endUsage);
-        
+
 
     #endregion
 
@@ -96,7 +96,7 @@ public interface IECGridOSClient
 
     #region Interchange
 
-    Task<bool> InterchangeCancel(string sessionID, int interchangeID);
+    Task<bool> InterchangeCancel(string sessionID, long interchangeID);
     Task<DateTime> InterchangeDate(string sessionID, string interchangeHeader);
     Task<InterchangeIDInfo> InterchangeHeaderInfo(string sessionID, string interchangeHeader);
     Task<InterchangeIDInfo> InterchangeHeaderInfoB(string sessionID, byte[] interchangeHeader);
@@ -137,15 +137,15 @@ public interface IECGridOSClient
     Task<List<InterconnectIDInfo>> InterconnectListByStatusEx(string sessionID, int networkID, int mailboxID, StatusInterconnect intStatus, int eCGridID, short maxDays);
     Task<bool> InterconnectNote(string sessionID, int interconnectID, AuthLevel authLevel, string note, string attachmentName, byte[] attachmentContent, string eMailTo, string otherEMailAddress);
     Task<List<InterconnectNote>> InterconnectNoteList(string sessionID, int interconnectID);
-        
+
     #endregion
 
     #region Key Set/Get/List/Remove
 
-    Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared);
-    Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, int objectID);
-    Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared);
-    Task<bool> KeySet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility, string value, string meta, int daysToLive);
+    Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared);
+    Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, long objectID);
+    Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared);
+    Task<bool> KeySet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility, string value, string meta, int daysToLive);
 
     #endregion
 
@@ -242,15 +242,15 @@ public interface IECGridOSClient
     Task<ParcelIDInfoCollection> ParcelOutBoxInProcess(string sessionID);
     Task<ParcelIDInfoCollection> ParcelOutBoxInProcessEx(string sessionID, int networkID, int mailboxID);
     Task<bool> ParcelSetMailbagControlID(string sessionID, long parcelID, string mailbagControlID);
-    Task<int> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType);
+    Task<long> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType);
     Task<bool> ParcelUpdateLocalStatus(string sessionID, long parcelID, short status);
     Task<bool> ParcelUpdateStatus(string sessionID, long parcelID, ParcelStatus status, bool transLogOnly);
-    Task<int> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content);
-    Task<int> ParcelUploadA(string sessionID, string fileName, string contentBase64);
-    Task<int> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content);
-    Task<int> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64);
-    Task<int> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo);
-    Task<int> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo);
+    Task<long> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content);
+    Task<long> ParcelUploadA(string sessionID, string fileName, string contentBase64);
+    Task<long> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content);
+    Task<long> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64);
+    Task<long> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo);
+    Task<long> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo);
 
     #endregion
 
@@ -772,7 +772,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, int callBackQueueID)
+    public async Task<CallBackQueueIDInfo> CallBackQueueInfo(string sessionID, long callBackQueueID)
     {
         try
         {
@@ -813,7 +813,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<CallBackQueueIDInfo> CallBackTest(string sessionID, int callBackQueueID, int parcelID, int interchangeID, int userID)
+    public async Task<CallBackQueueIDInfo> CallBackTest(string sessionID, long callBackQueueID, long parcelID, long interchangeID, int userID)
     {
         try
         {
@@ -2160,7 +2160,7 @@ public class ECGridOSClient : IECGridOSClient
 
     #region Interchange
 
-    public async Task<bool> InterchangeCancel(string sessionID, int interchangeID)
+    public async Task<bool> InterchangeCancel(string sessionID, long interchangeID)
     {
         try
         {
@@ -3029,7 +3029,7 @@ public class ECGridOSClient : IECGridOSClient
             if (responseHeader.IsSuccessStatusCode)
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
-                using StreamReader stream_reader = new StreamReader(responseStream); 
+                using StreamReader stream_reader = new StreamReader(responseStream);
                 var serializer = new XmlSerializer(typeof(List<InterchangeIDInfo>), domain_name);
                 return (List<InterchangeIDInfo>)serializer.Deserialize(stream_reader);
             }
@@ -3073,7 +3073,7 @@ public class ECGridOSClient : IECGridOSClient
             if (responseHeader.IsSuccessStatusCode)
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
-                using StreamReader stream_reader = new StreamReader(responseStream); 
+                using StreamReader stream_reader = new StreamReader(responseStream);
                 var serializer = new XmlSerializer(typeof(List<InterchangeIDInfo>), domain_name);
                 return (List<InterchangeIDInfo>)serializer.Deserialize(stream_reader);
             }
@@ -3760,7 +3760,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="objectID"></param>
     /// <param name="visibility"></param>
     /// <returns>KeyValue</returns>
-    public async Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared)
+    public async Task<KeyValue> KeyGet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared)
     {
         try
         {
@@ -3811,7 +3811,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="systemObject"></param>
     /// <param name="objectID"></param>
     /// <returns>List of KeyValue Objects</returns>
-    public async Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, int objectID)
+    public async Task<List<KeyValue>> KeyList(string sessionID, Objects systemObject, long objectID)
     {
         try
         {
@@ -3862,7 +3862,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="objectID"></param>
     /// <param name="visibility"></param>
     /// <returns>True/False</returns>
-    public async Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility = KeyVisibility.Shared)
+    public async Task<bool> KeyRemove(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility = KeyVisibility.Shared)
     {
         try
         {
@@ -3918,7 +3918,7 @@ public class ECGridOSClient : IECGridOSClient
     /// <param name="meta"></param>
     /// <param name="daysToLive"></param>
     /// <returns>True/False</returns>
-    public async Task<bool> KeySet(string sessionID, string key, Objects systemObject, int objectID, KeyVisibility visibility, string value, string meta, int daysToLive)
+    public async Task<bool> KeySet(string sessionID, string key, Objects systemObject, long objectID, KeyVisibility visibility, string value, string meta, int daysToLive)
     {
         try
         {
@@ -6900,7 +6900,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType)
+    public async Task<long> ParcelTest(string sessionID, int eCGridIDFrom, int eCGridIDTo, EDIStandard documentType)
     {
         try
         {
@@ -6925,8 +6925,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7028,7 +7028,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content)
+    public async Task<long> ParcelUpload(string sessionID, string fileName, int bytes, byte[] content)
     {
         try
         {
@@ -7053,8 +7053,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7071,7 +7071,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadA(string sessionID, string fileName, string contentBase64)
+    public async Task<long> ParcelUploadA(string sessionID, string fileName, string contentBase64)
     {
         try
         {
@@ -7095,8 +7095,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7113,7 +7113,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content)
+    public async Task<long> ParcelUploadEx(string sessionID, int networkID, int mailboxID, string fileName, int bytes, byte[] content)
     {
         try
         {
@@ -7140,8 +7140,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7158,7 +7158,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64)
+    public async Task<long> ParcelUploadExA(string sessionID, int networkID, int mailboxID, string fileName, string contentBase64)
     {
         try
         {
@@ -7184,8 +7184,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7202,7 +7202,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo)
+    public async Task<long> ParcelUploadMft(string sessionID, string fileName, int bytes, byte[] content, int eCGridIDFrom, int eCGridIDTo)
     {
         try
         {
@@ -7229,8 +7229,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
@@ -7247,7 +7247,7 @@ public class ECGridOSClient : IECGridOSClient
         }
     }
 
-    public async Task<int> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo)
+    public async Task<long> ParcelUploadMftA(string sessionID, string fileName, string contentBase64, int eCGridIDFrom, int eCGridIDTo)
     {
         try
         {
@@ -7273,8 +7273,8 @@ public class ECGridOSClient : IECGridOSClient
             {
                 using var responseStream = await responseHeader.Content.ReadAsStreamAsync();
                 using StreamReader stream_reader = new StreamReader(responseStream);
-                var serializer = new XmlSerializer(typeof(int), domain_name);
-                return (int)serializer.Deserialize(stream_reader);
+                var serializer = new XmlSerializer(typeof(long), domain_name);
+                return (long)serializer.Deserialize(stream_reader);
             }
             else
             {
